Remove /var/lib/NetworkManager from list of filesystems.  I see errors unmounting it during shutdown.  Add root homedir.
Allow for restarting after a failure: if we see the debootstrap snapshot on the
dest pool, assume that some script later than mkzpool failed and the user is
retrying after changing something.  Once that snapshot exists, user must
destroy dest pool if they want to rerun mkzpool on it.  We now destroy the pool
on exit if mkzpool created it and encounters a failure subsequently, so that
the user doesn't have to.